# src/docs/swagger-output.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "title": "Ciné Book Scrapping REST API",
    "description": "",
    "license": {
      "name": "MIT Licence",
      "url": "https://github.com/NemesisX1/canal_olympia_scrapping_api/blob/main/LICENSE.md"
    }
  },
...
  "securityDefinitions": {},
  "paths": {
...
    "/movies": {
      "get": {
        "tags": [
          "Movies"
        ],
        "summary": "Fetch all the available movies",
        "description": "Fetch all the available movies",
        "parameters": [
          {
            "name": "lang",
            "in": "query",
            "description": "The response language",
            "schema": {
              "type": "string",
              "enum": [
                "fr",
                "en"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "title": {
                      "type": "string",
                      "example": "The creator"
                    },
                    "url": {
                      "type": "string",
                      "example": "https://example.com"
                    },
                    "img": {
                      "type": "string",
                      "example": "https://example.com"
                    },
                    "slug": {
                      "type": "string",
                      "example": "the-creator"
                    }
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              },
              "application/xml": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "title": {
                      "type": "string",
                      "example": "The creator"
                    },
                    "url": {
                      "type": "string",
                      "example": "https://example.com"
                    },
                    "img": {
                      "type": "string",
                      "example": "https://example.com"
                    },
                    "slug": {
                      "type": "string",
                      "example": "the-creator"
                    }
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Bad request"
                    },
                    "errors": {
                      "type": "string",
                      "example": "any"
                    }
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              },
              "application/xml": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Bad request"
                    },
                    "errors": {
                      "type": "string",
                      "example": "any"
                    }
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              }
            }
          },
          "422": {
            "description": "Your body or query was bad formatted",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Your body or query was bad formatted"
                    },
                    "errors": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "type": {
                            "type": "string",
                            "example": "field"
                          },
                          "value": {
                            "type": "string",
                            "example": "zk"
                          },
                          "msg": {
                            "type": "string",
                            "example": "Invalid value"
                          },
                          "path": {
                            "type": "string",
                            "example": "lang"
                          },
                          "location": {
                            "type": "string",
                            "example": "query"
                          }
                        }
                      }
                    }
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              },
              "application/xml": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Your body or query was bad formatted"
                    },
                    "errors": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "type": {
                            "type": "string",
                            "example": "field"
                          },
                          "value": {
                            "type": "string",
                            "example": "zk"
                          },
                          "msg": {
                            "type": "string",
                            "example": "Invalid value"
                          },
                          "path": {
                            "type": "string",
                            "example": "lang"
                          },
                          "location": {
                            "type": "string",
                            "example": "query"
                          }
                        }
                      }
                    }
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/movies/{theaterSlug}": {
      "get": {
        "tags": [
          "Movies"
        ],
        "summary": "Fetch all the available movies at a given theater",
        "description": "Fetch all the available movies at a given theater",
        "parameters": [
          {
            "name": "theaterSlug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "lang",
            "in": "query",
            "description": "The response language",
            "schema": {
              "type": "string",
              "enum": [
                "fr",
                "en"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TheaterMovieBrief"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/TheaterMovieBrief"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Bad request"
                    },
                    "errors": {
                      "type": "string",
                      "example": "any"
                    }
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              },
              "application/xml": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Bad request"
                    },
                    "errors": {
                      "type": "string",
                      "example": "any"
                    }
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Theater not found"
                    },
                    "errors": {
                      "type": "array",
                      "example": [],
                      "items": {}
                    }
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              },
              "application/xml": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Theater not found"
                    },
                    "errors": {
                      "type": "array",
                      "example": [],
                      "items": {}
                    }
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              }
            }
          },
          "422": {
            "description": "Your body or query was bad formatted",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Your body or query was bad formatted"
                    },
                    "errors": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "type": {
                            "type": "string",
                            "example": "field"
                          },
                          "value": {
                            "type": "string",
                            "example": "zk"
                          },
                          "msg": {
                            "type": "string",
                            "example": "Invalid value"
                          },
                          "path": {
                            "type": "string",
                            "example": "lang"
                          },
                          "location": {
                            "type": "string",
                            "example": "query"
                          }
                        }
                      }
                    }
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              },
              "application/xml": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Your body or query was bad formatted"
                    },
                    "errors": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "type": {
                            "type": "string",
                            "example": "field"
                          },
                          "value": {
                            "type": "string",
                            "example": "zk"
                          },
                          "msg": {
                            "type": "string",
                            "example": "Invalid value"
                          },
                          "path": {
                            "type": "string",
                            "example": "lang"
                          },
                          "location": {
                            "type": "string",
                            "example": "query"
                          }
                        }
                      }
                    }
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/movies/infos/{slug}": {
      "get": {
        "tags": [
          "Movies"
        ],
        "summary": "Fetch movie informations for a given movie slug",
        "description": "Fetch movie informations for a given movie slug",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "lang",
            "in": "query",
            "description": "The response language",
            "schema": {
              "type": "string",
              "enum": [
                "fr",
                "en"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TheaterMovie"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/TheaterMovie"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Bad request"
                    },
                    "errors": {
                      "type": "string",
                      "example": "any"
                    }
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              },
              "application/xml": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Bad request"
                    },
                    "errors": {
                      "type": "string",
                      "example": "any"
                    }
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "this slug was not found"
                    },
                    "errors": {
                      "type": "array",
                      "example": [],
                      "items": {}
                    }
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              },
              "application/xml": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "this slug was not found"
                    },
                    "errors": {
                      "type": "array",
                      "example": [],
                      "items": {}
                    }
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              }
            }
          },
          "422": {
            "description": "Your body or query was bad formatted",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Your body or query was bad formatted"
                    },
                    "errors": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "type": {
                            "type": "string",
                            "example": "field"
                          },
                          "value": {
                            "type": "string",
                            "example": "zk"
                          },
                          "msg": {
                            "type": "string",
                            "example": "Invalid value"
                          },
                          "path": {
                            "type": "string",
                            "example": "lang"
                          },
                          "location": {
                            "type": "string",
                            "example": "query"
                          }
                        }
                      }
                    }
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              },
              "application/xml": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Your body or query was bad formatted"
                    },
                    "errors": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "type": {
                            "type": "string",
                            "example": "field"
                          },
                          "value": {
                            "type": "string",
                            "example": "zk"
                          },
                          "msg": {
                            "type": "string",
                            "example": "Invalid value"
                          },
                          "path": {
                            "type": "string",
                            "example": "lang"
                          },
                          "location": {
                            "type": "string",
                            "example": "query"
                          }
                        }
                      }
                    }
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "TheaterMovie": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "example": "The Creator"
          },
          "slug": {
            "type": "string",
            "example": "the-creator"
          },
          "genre": {
            "type": "string",
            "example": "Horror"
          },
          "duration": {
            "type": "string",
            "example": "1h20"
          },
          "releaseDate": {
            "type": "string",
            "example": "10-10-2023"
          },
          "descriptionBrief": {
            "type": "string",
            "example": "This is a nice movie"
          },
          "trailerLink": {
            "type": "string",
            "example": "https://example.com"
          }
        },
        "xml": {
          "name": "TheaterMovie"
        }
      },
      "TheaterMovieBrief": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "example": "The Creator"
          },
          "slug": {
            "type": "string",
            "example": "the-creator"
          },
          "date": {
            "type": "string",
            "example": "10-10-2023"
          },
          "hour": {
            "type": "string",
            "example": "1h12"
          },
          "language": {
            "type": "string",
            "example": "en"
          },
          "img": {
            "type": "string",
            "example": "https://example.com"
          },
          "url": {
            "type": "string",
            "example": "https://example.com"
          }
        },
        "xml": {
          "name": "TheaterMovieBrief"
        }
      },
...
    }
  },
...
}